feat: broke out EarnNetworkAvatar into its own file
