+++ app/components/UI/Earn/components/EarnNetworkAvatar/EarnNetworkAvatar.styles.ts
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+
+const styleSheet = () =>
+  StyleSheet.create({
+    networkAvatar: {
+      height: 32,
+      width: 32,
+      flexShrink: 0,
+    },
+  });
+
+export default styleSheet;
